Centralize Cypress authentication stubs
