+++ cypress.config.ts
@@ -32,6 +32,6 @@ export default defineConfig({
   e2e: {
     baseUrl: 'http://localhost:4200',
     specPattern: 'cypress/e2e/**/*.cy.ts',
-    supportFile: false
+    supportFile: 'cypress/support/e2e.ts'
   }
 });
